# openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 1.0.0
  description: API for creating, managing, and resolving short links.
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8080
security: []
paths:
  /api/v1/links:
    post:
      summary: Create a short link
      operationId: createLink
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkRequest'
      responses:
        '201':
          description: Short link created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/links/{code}:
    parameters:
      - $ref: '#/components/parameters/Code'
    get:
      summary: Get link metadata
      operationId: getLink
      responses:
        '200':
          description: Link metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
    delete:
      summary: Disable a short link
      operationId: disableLink
      responses:
        '204':
          description: Link disabled.
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
  /api/v1/links/{code}/analytics:
    parameters:
      - $ref: '#/components/parameters/Code'
    get:
      summary: Get click analytics for a short link
      description: >-
        Returns analytics for the half-open interval [from, to). When neither parameter is
        provided, the interval defaults to the trailing 30 days ending at the current time.
        A requested interval cannot exceed 90 days. Analytics remains accessible for expired
        or disabled links.
      operationId: getLinkAnalytics
      parameters:
        - name: from
          in: query
          required: false
          description: ISO-8601 interval start. Defaults to 30 days before `to`.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: ISO-8601 exclusive interval end. Defaults to the current time.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Click analytics for the requested interval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          $ref: '#/components/responses/InvalidAnalyticsRange'
        '404':
          $ref: '#/components/responses/NotFound'
  /{code}:
    parameters:
      - $ref: '#/components/parameters/Code'
    get:
      summary: Redirect to the original URL
      operationId: redirectToLink
      responses:
        '302':
          description: Redirect to the active destination.
          headers:
            Location:
              description: Original destination URL.
              required: true
              schema:
                type: string
                format: uri
            Cache-Control:
              description: Prevents caching of redirects.
              required: true
              schema:
                type: string
                enum: [no-store]
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
components:
  parameters:
    Code:
      name: code
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  schemas:
    CreateLinkRequest:
      type: object
      required: [originalUrl]
      properties:
        originalUrl:
          type: string
          format: uri
          maxLength: 2048
          pattern: '^https?://.+'
          description: Absolute HTTP or HTTPS destination URL.
        customAlias:
          type: string
          minLength: 4
          maxLength: 32
          pattern: '^[A-Za-z0-9_-]+$'
        expiresAt:
          type: string
          format: date-time
          description: Future ISO-8601 timestamp.
    LinkResponse:
      type: object
      required: [code, shortUrl, originalUrl, createdAt, expiresAt]
      properties:
        code:
          type: string
        shortUrl:
          type: string
          format: uri
        originalUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    AnalyticsResponse:
      type: object
      required: [code, from, to, totalClicks, topReferrers]
      properties:
        code:
          type: string
          description: Short link code.
        from:
          type: string
          format: date-time
          description: Inclusive start of the analytics interval.
        to:
          type: string
          format: date-time
          description: Exclusive end of the analytics interval.
        totalClicks:
          type: integer
          format: int64
          minimum: 0
        topReferrers:
          type: array
          items:
            $ref: '#/components/schemas/ReferrerStat'
    ReferrerStat:
      type: object
      required: [referrer, clicks]
      properties:
        referrer:
          type: string
          description: Referrer hostname, or `direct` when no usable referrer was supplied.
        clicks:
          type: integer
          format: int64
          minimum: 0
    ErrorResponse:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InvalidAnalyticsRange:
      description: Invalid analytics range. The error response code is `INVALID_ANALYTICS_RANGE`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Alias is unavailable or the idempotency key conflicts with a prior request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Link not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Gone:
      description: Link is expired or disabled.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Code generation failed after repeated collisions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'